Vendor textgrid.py from NLTK.
We only need this one file, so it seems overly restrictive to depend on
the entire nltk-contrib project.  Instead I've copied in just the file
we need, and added a script that will check to make sure our version
matches the one from the nltk-contrib repo.  Eventually I'll call that
script from the script that builds releases.