Emulate the matlab wavread function.
The algorithms expect matlab format, so convert the wav data to it.  I
wrote this as an emulation of the 'wavread' calling API, since that will
make it easier to compare the matlab and python code.  Eventually the
two will diverge, at which point it may become desirable to refactor
this.