Use scipy.io.loadmat, and actually commit the data.
I forgot to commit the sample data in the first commit.  The actual
checked in data here is in matlab binary format, so we can load it using
scip.io.loadmat.  That allows us to to put all the expected data in the
data file.  I've also had second thoughts about whether the floats
will compare equal on all platforms, so I've changed the doc note
to explain how to fix it if it turns out they don't.