Clean up error messages, more settings tests.
I dropped CmdError in favor of using the argparse error function, but
I don't like the way it prints out the usage before the error message
so I subclassed it and cleaned that up.  That also allowed me to
fix an argparse bug where it would display a useless name (__main__.py)
in error messages when the script is run via -m.

In the process of adding a test that measurements can be specified in
the settings file, I added a check (and test) that you have to specify
at least one measurement.