Fix tests of line length, handling of include_F0.
Somehow I managed to write assertTrue instead of assertEqual, and of
course it passed since the second argument was taken as the message
and the first argument was always True.  Fixing this led to the
discover (a) that the code handling --inculde-F0 was buggy and
(b) that the way I was splitting lines in the tests was dropping
blank fields.

The fix for (a) is to drop the special handling of F0 in the
_assemble_fields function, since in fact I'm adding the field to
the list of data fields during argument handling.

The fix for (b) was to change the setup function to do the line
stripping and splitting on the delimiter.  This meant I had to
use the full filename in the multi-filename test since now the
assertIn is checking against elements of a list instead of
against a string.