Make output to stdout the default.
This makes the command act more like a normal unix command, and if
someone wants output to a file to be the default they can specify
it in their .opensaucerc file.

This also speeds up the tests slightly since we no longer need to
create a tempdir and clean it up for each test.